fix: CodeRabbit 리뷰 반영 (Place/Course id Long, forked_from_id FK, order_no NOT NULL, 주석 정정)

+++ src/main/kotlin/com/example/backend/place/domain/model/Place.kt
@@ -5,7 +5,7 @@ package com.example.backend.place.domain.model
  */
 @ConsistentCopyVisibility // copy() 도 private 으로 — 팩토리 우회 차단
 data class Place private constructor(
-    val id: Int?,
+    val id: Long?,
     val name: String,
 ) {
     companion object {
